Nettoyage du widget Contact grid view avec l'ajout de 3 fonctions séparées
