+++ lib/features/events/presentation/pages/event_show_page/pop_page/event_management_page.dart/sections/members_section/functions/show_message.dart
@@ -1,0 +1,8 @@
+import 'package:faro_clean_tdd/core/util/show_result_message_snackbar.dart';
+import 'package:flutter/material.dart';
+
+void showMessage(BuildContext context, String message) {
+  if (context.mounted) {
+    showResultMessageSnackbar(context: context, message: message);
+  }
+}
