+++ lib/features/events/presentation/pages/event_show_page/pop_page/event_management_page.dart/sections/members_section/widgets/add_member/components/contact_grid_view.dart
@@ -1,14 +1,10 @@
-import 'package:faro_clean_tdd/core/util/show_result_message_snackbar.dart';
 import 'package:faro_clean_tdd/features/contacts/domain/entities/contact.dart';
 import 'package:faro_clean_tdd/features/events/domain/entities/event.dart';
-import 'package:faro_clean_tdd/features/events/presentation/providers/fetch_event/fetch_event_provider.dart';
+import 'package:faro_clean_tdd/features/events/presentation/pages/event_show_page/pop_page/event_management_page.dart/sections/members_section/functions/show_create_member_result.dart';
+import 'package:faro_clean_tdd/features/events/presentation/pages/event_show_page/pop_page/event_management_page.dart/sections/members_section/functions/show_update_ticket_result.dart';
 import 'package:faro_clean_tdd/features/members/data/models/member_model.dart';
 import 'package:faro_clean_tdd/features/members/presentation/providers/create_member/create_member_provider.dart';
-import 'package:faro_clean_tdd/features/members/presentation/providers/create_member/state/create_member_state.dart';
 import 'package:faro_clean_tdd/features/tickets/domain/entities/ticket.dart';
-import 'package:faro_clean_tdd/features/tickets/presentation/providers/fetch_tickets/fetch_tickets_provider.dart';
-import 'package:faro_clean_tdd/features/tickets/presentation/providers/update_ticket/state/update_ticket_state.dart'
-    as ut;
 import 'package:faro_clean_tdd/features/tickets/presentation/providers/update_ticket/update_ticket_provider.dart';
 import 'package:faro_clean_tdd/features/events/presentation/pages/event_show_page/pop_page/event_management_page.dart/sections/members_section/widgets/add_member/components/contact_card.dart';
 import 'package:flutter/material.dart';
@@ -29,54 +25,6 @@ class ContactGridView extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    void showMessage(BuildContext context, String message) {
-      if (context.mounted) {
-        showResultMessageSnackbar(context: context, message: message);
-      }
-    }
-
-    void showResultCreateMember(
-      BuildContext context,
-      CreateMemberState state,
-    ) {
-      final bool isSucess = state is Loaded && context.mounted;
-      final bool isError = state is Error && context.mounted;
-
-      if (isSucess) {
-        final fetchEventState = ref.read(fetchEventProvider);
-        ref.read(fetchEventProvider.notifier).addMemberToEvent(
-              member: state.member,
-              fetchEventState: fetchEventState,
-              event: event!,
-            );
-        showMessage(context, state.message);
-      }
-
-      if (isError) {
-        showMessage(context, state.message);
-      }
-    }
-
-    void showResultUpdateTicket(
-      BuildContext context,
-      ut.UpdateTicketState state,
-    ) {
-      final bool isSucess = state is ut.Loaded && context.mounted;
-      final bool isError = state is ut.Error && context.mounted;
-
-      if (isSucess) {
-        final fetchTicketState = ref.read(fetchTicketsProvider);
-        ref.read(fetchTicketsProvider.notifier).removeTicket(
-            fetchTicketsState: fetchTicketState, ticket: state.ticket);
-        showMessage(context, state.message);
-        Navigator.of(context).pop();
-      }
-
-      if (isError) {
-        showMessage(context, state.message);
-      }
-    }
-
     final bool isForAddMember = event != null;
     return SizedBox(
       height: mediaHeight * 0.2,
@@ -110,7 +58,12 @@ class ContactGridView extends ConsumerWidget {
                                 username: contact.username),
                           );
                       if (context.mounted) {
-                        showResultCreateMember(context, stateResult);
+                        showCreateMemberResult(
+                          context,
+                          stateResult,
+                          ref,
+                          event!,
+                        );
                       }
                     } else {
                       final stateResult = await ref
@@ -120,7 +73,11 @@ class ContactGridView extends ConsumerWidget {
                             userId: contact.userId,
                           );
                       if (context.mounted) {
-                        showResultUpdateTicket(context, stateResult);
+                        showUpdateTicketResult(
+                          context,
+                          stateResult,
+                          ref,
+                        );
                       }
                     }
                   },
